+++ readable-code/src/main/java/cleancode/studycafe/misson7/exception/AppException.java
@@ -1,0 +1,9 @@
+package cleancode.studycafe.misson7.exception;
+
+public class AppException extends RuntimeException {
+
+    public AppException(String message) {
+        super(message);
+    }
+
+}
