+++ readable-code/src/main/java/cleancode/studycafe/misson7/exception/ProvideException.java
@@ -1,0 +1,9 @@
+package cleancode.studycafe.misson7.exception;
+
+public class ProvideException extends RuntimeException {
+
+    public ProvideException(String message) {
+        super(message);
+    }
+
+}
