+++ readable-code/src/main/java/cleancode/studycafe/misson7/format/PassTypeFormatter.java
@@ -1,0 +1,8 @@
+package cleancode.studycafe.misson7.format;
+
+import cleancode.studycafe.misson7.model.pass.StudyCafePass;
+
+public interface PassTypeFormatter {
+
+    String format(StudyCafePass pass);
+}
