+++ readable-code/src/main/java/cleancode/studycafe/misson7/io/handler/ConsoleInputHandler.java
@@ -1,0 +1,32 @@
+package cleancode.studycafe.misson7.io.handler;
+
+import cleancode.studycafe.misson7.model.pass.StudyCafePassType;
+import cleancode.studycafe.misson7.model.pass.seat.StudyCafeSeatPass;
+import cleancode.studycafe.misson7.model.pass.seat.StudyCafeSeatPasses;
+
+import java.util.Scanner;
+
+public class ConsoleInputHandler implements InputHandler {
+
+    private static final Scanner SCANNER = new Scanner(System.in);
+
+    @Override
+    public StudyCafePassType getPassTypeSelectingUserAction() {
+        String userInput = SCANNER.nextLine();
+        return StudyCafePassType.findBy(userInput);
+    }
+
+    @Override
+    public StudyCafeSeatPass getSelectSeatPass(StudyCafeSeatPasses seatPasses) {
+        String userInput = SCANNER.nextLine();
+        int selectedIndex = Integer.parseInt(userInput) - 1;
+        return seatPasses.findBy(selectedIndex);
+    }
+
+    @Override
+    public boolean getLockerSelection() {
+        String userInput = SCANNER.nextLine();
+        return "1".equals(userInput);
+    }
+
+}
