+++ readable-code/src/main/java/cleancode/studycafe/misson7/io/handler/InputHandler.java
@@ -1,0 +1,15 @@
+package cleancode.studycafe.misson7.io.handler;
+
+import cleancode.studycafe.misson7.model.pass.StudyCafePassType;
+import cleancode.studycafe.misson7.model.pass.seat.StudyCafeSeatPass;
+import cleancode.studycafe.misson7.model.pass.seat.StudyCafeSeatPasses;
+
+public interface InputHandler {
+
+    StudyCafePassType getPassTypeSelectingUserAction();
+
+    StudyCafeSeatPass getSelectSeatPass(StudyCafeSeatPasses seatPasses);
+
+    boolean getLockerSelection();
+
+}
