+++ readable-code/src/main/java/cleancode/studycafe/misson7/model/pass/StudyCafePass.java
@@ -1,0 +1,10 @@
+package cleancode.studycafe.misson7.model.pass;
+
+public interface StudyCafePass {
+
+    StudyCafePassType getPassType();
+
+    int getDuration();
+
+    int getPrice();
+}
