+++ readable-code/src/main/java/cleancode/studycafe/misson7/model/pass/locker/StudyCafeLockerPass.java
@@ -1,0 +1,45 @@
+package cleancode.studycafe.misson7.model.pass.locker;
+
+import cleancode.studycafe.misson7.model.pass.StudyCafePass;
+import cleancode.studycafe.misson7.model.pass.StudyCafePassType;
+
+public class StudyCafeLockerPass implements StudyCafePass {
+
+    private final StudyCafePassType passType;
+    private final int duration;
+    private final int price;
+
+    private StudyCafeLockerPass(StudyCafePassType passType, int duration, int price) {
+        this.passType = passType;
+        this.duration = duration;
+        this.price = price;
+    }
+
+    public static StudyCafeLockerPass of(StudyCafePassType passType, int duration, int price) {
+        return new StudyCafeLockerPass(passType, duration, price);
+    }
+
+    public boolean isSamePassType(StudyCafePassType passType) {
+        return passType == this.passType;
+    }
+
+    public boolean isSameDuration(int duration) {
+        return duration == this.duration;
+    }
+
+    @Override
+    public StudyCafePassType getPassType() {
+        return passType;
+    }
+
+    @Override
+    public int getDuration() {
+        return duration;
+    }
+
+    @Override
+    public int getPrice() {
+        return price;
+    }
+
+}
