+++ readable-code/src/main/java/cleancode/studycafe/misson7/provider/LockerPassProvider.java
@@ -1,0 +1,8 @@
+package cleancode.studycafe.misson7.provider;
+
+import cleancode.studycafe.misson7.model.pass.locker.StudyCafeLockerPasses;
+
+public interface LockerPassProvider {
+
+    StudyCafeLockerPasses provide();
+}
