+++ readable-code/src/main/java/cleancode/studycafe/misson7/provider/SeatPassProvider.java
@@ -1,0 +1,8 @@
+package cleancode.studycafe.misson7.provider;
+
+import cleancode.studycafe.misson7.model.pass.seat.StudyCafeSeatPasses;
+
+public interface SeatPassProvider {
+
+    StudyCafeSeatPasses provide();
+}
