+++ readable-code/src/main/java/cleancode/studycafe/misson7/select/PassTypeSelectable.java
@@ -1,0 +1,6 @@
+package cleancode.studycafe.misson7.select;
+
+public interface PassTypeSelectable {
+
+    boolean selected(String userInput);
+}
