--- readable-code/src/main/java/cleancode/studycafe/tobe/exception/AppException.java
@@ -1,9 +1,0 @@
-package cleancode.studycafe.tobe.exception;
-
-public class AppException extends RuntimeException {
-
-    public AppException(String message) {
-        super(message);
-    }
-
-}
