--- readable-code/src/main/java/cleancode/studycafe/tobe/exception/ProvideException.java
@@ -1,9 +1,0 @@
-package cleancode.studycafe.tobe.exception;
-
-public class ProvideException extends RuntimeException {
-
-    public ProvideException(String message) {
-        super(message);
-    }
-
-}
