--- readable-code/src/main/java/cleancode/studycafe/tobe/format/PassTypeFormatter.java
@@ -1,8 +1,0 @@
-package cleancode.studycafe.tobe.format;
-
-import cleancode.studycafe.tobe.model.pass.StudyCafePass;
-
-public interface PassTypeFormatter {
-
-    String format(StudyCafePass pass);
-}
