--- readable-code/src/main/java/cleancode/studycafe/tobe/io/handler/InputHandler.java
@@ -1,15 +1,0 @@
-package cleancode.studycafe.tobe.io.handler;
-
-import cleancode.studycafe.tobe.model.pass.StudyCafePassType;
-import cleancode.studycafe.tobe.model.pass.seat.StudyCafeSeatPass;
-import cleancode.studycafe.tobe.model.pass.seat.StudyCafeSeatPasses;
-
-public interface InputHandler {
-
-    StudyCafePassType getPassTypeSelectingUserAction();
-
-    StudyCafeSeatPass getSelectSeatPass(StudyCafeSeatPasses seatPasses);
-
-    boolean getLockerSelection();
-
-}
