--- readable-code/src/main/java/cleancode/studycafe/tobe/model/pass/StudyCafePass.java
@@ -1,10 +1,0 @@
-package cleancode.studycafe.tobe.model.pass;
-
-public interface StudyCafePass {
-
-    StudyCafePassType getPassType();
-
-    int getDuration();
-
-    int getPrice();
-}
