--- readable-code/src/main/java/cleancode/studycafe/tobe/model/pass/seat/StudyCafeSeatPass.java
@@ -1,51 +1,0 @@
-package cleancode.studycafe.tobe.model.pass.seat;
-
-import cleancode.studycafe.tobe.model.pass.StudyCafePass;
-import cleancode.studycafe.tobe.model.pass.StudyCafePassType;
-import cleancode.studycafe.tobe.model.pass.locker.StudyCafeLockerPass;
-
-public class StudyCafeSeatPass implements StudyCafePass {
-
-    private final StudyCafePassType passType;
-    private final int duration;
-    private final int price;
-    private final double discountRate;
-
-    private StudyCafeSeatPass(StudyCafePassType passType, int duration, int price, double discountRate) {
-        this.passType = passType;
-        this.duration = duration;
-        this.price = price;
-        this.discountRate = discountRate;
-    }
-
-    public static StudyCafeSeatPass of(StudyCafePassType passType, int duration, int price, double discountRate) {
-        return new StudyCafeSeatPass(passType, duration, price, discountRate);
-    }
-
-    public boolean cannotSelectLockerPass() {
-        return passType != StudyCafePassType.FIXED;
-    }
-
-    public boolean isCandidate(StudyCafeLockerPass lockerPass) {
-        return lockerPass.isSamePassType(passType) && lockerPass.isSameDuration(duration);
-    }
-
-    @Override
-    public StudyCafePassType getPassType() {
-        return passType;
-    }
-
-    @Override
-    public int getDuration() {
-        return duration;
-    }
-
-    @Override
-    public int getPrice() {
-        return price;
-    }
-
-    public int getDiscountPrice() {
-        return (int) (price * discountRate);
-    }
-}
