--- readable-code/src/main/java/cleancode/studycafe/tobe/model/pass/seat/StudyCafeSeatPasses.java
@@ -1,34 +1,0 @@
-package cleancode.studycafe.tobe.model.pass.seat;
-
-import cleancode.studycafe.tobe.model.pass.StudyCafePassType;
-
-import java.util.List;
-
-public class StudyCafeSeatPasses {
-
-    private final List<StudyCafeSeatPass> passes;
-
-    private StudyCafeSeatPasses(List<StudyCafeSeatPass> passes) {
-        this.passes = passes;
-    }
-
-    public static StudyCafeSeatPasses of(List<StudyCafeSeatPass> passes) {
-        return new StudyCafeSeatPasses(passes);
-    }
-
-    public StudyCafeSeatPasses findCandidatesBy(StudyCafePassType passType) {
-        List<StudyCafeSeatPass> passCandidates = passes.stream()
-            .filter(passType::isSamePassType)
-            .toList();
-
-        return of(passCandidates);
-    }
-
-    public int size() {
-        return passes.size();
-    }
-
-    public StudyCafeSeatPass findBy(int index) {
-        return passes.get(index);
-    }
-}
