--- readable-code/src/main/java/cleancode/studycafe/tobe/provider/LockerPassProvider.java
@@ -1,8 +1,0 @@
-package cleancode.studycafe.tobe.provider;
-
-import cleancode.studycafe.tobe.model.pass.locker.StudyCafeLockerPasses;
-
-public interface LockerPassProvider {
-
-    StudyCafeLockerPasses provide();
-}
