--- readable-code/src/main/java/cleancode/studycafe/tobe/provider/SeatPassProvider.java
@@ -1,8 +1,0 @@
-package cleancode.studycafe.tobe.provider;
-
-import cleancode.studycafe.tobe.model.pass.seat.StudyCafeSeatPasses;
-
-public interface SeatPassProvider {
-
-    StudyCafeSeatPasses provide();
-}
