--- readable-code/src/main/java/cleancode/studycafe/tobe/select/PassTypeSelectable.java
@@ -1,6 +1,0 @@
-package cleancode.studycafe.tobe.select;
-
-public interface PassTypeSelectable {
-
-    boolean selected(String userInput);
-}
